Moved server code out.
